requirements.txt
beautifulsoup4==4.10.0
colorama==0.3.7
fastapi==0.70.0
inquirer2-windows==1.0.0
lxml==4.6.4
oauth==1.0.1
prompt-toolkit==3.0.22
psycopg2==2.9.1
Pygments==2.10.0
python-dotenv==0.19.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
urllib3==1.22
uvicorn==0.15.0
wadllib==1.3.2
watchgod==0.7
wikipedia==1.4.0
